requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=/home/vatche/.work/langchain/requirements.txt /home/vatche/.work/langchain/requirements.in
#
aiohttp==3.8.4
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via langchain
async-timeout==4.0.2
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
dataclasses-json==0.6.1
    # via langchain
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.1
    # via sqlalchemy
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.325
    # via -r /home/vatche/.work/langchain/requirements.in
langsmith==0.0.53
    # via langchain
loguru==0.7.0
    # via -r /home/vatche/.work/langchain/requirements.in
marshmallow==3.20.1
    # via dataclasses-json
mccabe==0.7.0
    # via pylama
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.25.1
    # via
    #   langchain
    #   pandas
openai==0.28.1
    # via -r /home/vatche/.work/langchain/requirements.in
packaging==23.1
    # via
    #   marshmallow
    #   pytest
pandas==2.0.3
    # via -r /home/vatche/.work/langchain/requirements.in
pluggy==1.2.0
    # via pytest
pycodestyle==2.11.1
    # via pylama
pydantic==2.4.2
    # via
    #   langchain
    #   langsmith
pydantic-core==2.10.1
    # via pydantic
pydocstyle==6.3.0
    # via pylama
pyflakes==3.1.0
    # via pylama
pylama==8.4.1
    # via -r /home/vatche/.work/langchain/requirements.in
pytest==7.4.0
    # via -r /home/vatche/.work/langchain/requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r /home/vatche/.work/langchain/requirements.in
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via langchain
requests==2.31.0
    # via
    #   -r /home/vatche/.work/langchain/requirements.in
    #   langchain
    #   langsmith
    #   openai
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
snowballstemmer==2.2.0
    # via pydocstyle
sqlalchemy==2.0.22
    # via langchain
tenacity==8.2.3
    # via langchain
tomli==2.0.1
    # via pytest
tqdm==4.66.1
    # via openai
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
urllib3==2.0.3
    # via requests
yarl==1.9.2
    # via aiohttp
